# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chat API",
    "version": "1.0.0"
  },
  "paths": {
    "/chats/": {
      "post": {
        "summary": "Create chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{id}/messages/": {
      "post": {
        "summary": "Send message to chat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
    "/chats/{id}": {
      "get": {
        "summary": "Get chat and last N messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat and messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatWithMessages"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found"
          }
        }
      },
      "delete": {
        "summary": "Delete chat and all messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Chat": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "title", "created_at"]
      },
      "ChatCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": ["title"]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "chat_id": { "type": "integer" },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "chat_id", "text", "created_at"]
      },
      "MessageCreateRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          }
        },
        "required": ["text"]
      },
      "ChatWithMessages": {
        "type": "object",
        "properties": {
          "chat": { "$ref": "#/components/schemas/Chat" },
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Message" }
          }
        },
        "required": ["chat", "messages"]
      }
    }
  }
}